Dock improvements
- Display small dock for screens up to 856px
- Display small icons on small dock
- Adapt font-size and padding on expanded small navbar
- Refactor smallNavbar mechanism to be in a directive + fix `onMatchMediaChange` binding
